a separated colorpicker component
